Fix and add
Replacing all use of np.column_stack with directing writing columns into
an empty array, because the former has the danger to chop data if they
are too long.
Modifying functions related to lambda for iISD to reflect the switch
from scaled to original parameter.
Adding functions to plot randomization results (to be further modified).